+++ cedar-policy-cli/src/lib.rs
@@ -21,7 +21,7 @@
 
 mod err;
 
-use clap::{Args, Parser, Subcommand, ValueEnum};
+use clap::{ArgAction, Args, Parser, Subcommand, ValueEnum};
 use miette::{miette, IntoDiagnostic, NamedSource, Report, Result, WrapErr};
 use serde::{Deserialize, Serialize};
 use std::{
@@ -139,10 +139,19 @@ pub struct RequestArgs {
     /// --principal, --action, etc.
     #[arg(long = "request-json", value_name = "FILE", conflicts_with_all = &["principal", "action", "resource", "context_json_file"])]
     pub request_json_file: Option<String>,
+    /// Whether to enable request validation. This has no effect if a schema is
+    /// not provided.
+    #[arg(long = "request-validation", action = ArgAction::Set, default_value_t = true)]
+    pub request_validation: bool,
 }
 
 impl RequestArgs {
     /// Turn this `RequestArgs` into the appropriate `Request` object
+    ///
+    /// `schema` will be used for schema-based parsing of the context, and also
+    /// (if `self.request_validation` is `true`) for request validation.
+    ///
+    /// `self.request_validation` has no effect if `schema` is `None`.
     fn get_request(&self, schema: Option<&Schema>) -> Result<Request> {
         match &self.request_json_file {
             Some(jsonfile) => {
@@ -182,7 +191,18 @@ impl RequestArgs {
                 )
                 .into_diagnostic()
                 .wrap_err_with(|| format!("failed to create a context from {jsonfile}"))?;
-                Ok(Request::new(principal, action, resource, context))
+                Request::new(
+                    principal,
+                    action,
+                    resource,
+                    context,
+                    if self.request_validation {
+                        schema
+                    } else {
+                        None
+                    },
+                )
+                .map_err(|e| miette!("{e}"))
             }
             None => {
                 let principal = self
@@ -224,7 +244,18 @@ impl RequestArgs {
                         })?,
                     },
                 };
-                Ok(Request::new(principal, action, resource, context))
+                Request::new(
+                    principal,
+                    action,
+                    resource,
+                    context,
+                    if self.request_validation {
+                        schema
+                    } else {
+                        None
+                    },
+                )
+                .map_err(|e| miette!("{e}"))
             }
         }
     }