+++ cedar-policy/src/api.rs
@@ -620,7 +620,7 @@ impl Authorizer {
     /// #
     /// # let c = Context::empty();
     /// #
-    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), c);
+    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), c, None).unwrap();
     /// #
     /// # // create a policy
     /// # let s = r#"permit(
@@ -676,7 +676,7 @@ impl Authorizer {
     ///
     /// let c = Context::empty();
     ///
-    /// let request: Request = Request::new(Some(p), Some(a), Some(r), c);
+    /// let request: Request = Request::new(Some(p), Some(a), Some(r), c, None).unwrap();
     ///
     /// // create a policy
     /// let s = r#"
@@ -802,7 +802,7 @@ impl Diagnostics {
     /// #
     /// # let c = Context::empty();
     /// #
-    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), c);
+    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), c, None).unwrap();
     /// #
     /// # // create a policy
     /// # let s = r#"permit(
@@ -861,7 +861,7 @@ impl Diagnostics {
     /// #
     /// # let c = Context::empty();
     /// #
-    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), c);
+    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), c, None).unwrap();
     /// #
     /// # // create a policy
     /// # let s = r#"permit(
@@ -2973,22 +2973,14 @@ impl<'a> RequestBuilder<'a> {
 
     /// Create the [`Request`]
     pub fn build(self) -> Result<Request, RequestValidationError> {
-        match self.schema {
-            None => Ok(Request(ast::Request::new_with_unknowns(
-                self.principal,
-                self.action,
-                self.resource,
-                self.context,
-            ))),
-            Some(schema) => Ok(Request(ast::Request::new_with_unknowns_and_validation(
-                self.principal,
-                self.action,
-                self.resource,
-                self.context,
-                &cedar_policy_validator::CoreSchema::new(&schema.0),
-                Extensions::all_available(),
-            )?)),
-        }
+        Ok(Request(ast::Request::new_with_unknowns(
+            self.principal,
+            self.action,
+            self.resource,
+            self.context,
+            self.schema.map(|schema| &schema.0),
+            Extensions::all_available(),
+        )?))
     }
 }
 
@@ -3013,42 +3005,15 @@ impl Request {
     /// decisions (e.g., because they are not used in your policies).
     /// If any of the fields are `None`, we will automatically generate
     /// a unique entity UID that is not equal to any UID in the store.
+    ///
+    /// If `schema` is present, this constructor will validate that the
+    /// `Request` complies with the given `schema`.
     pub fn new(
         principal: Option<EntityUid>,
         action: Option<EntityUid>,
         resource: Option<EntityUid>,
         context: Context,
-    ) -> Self {
-        let p = match principal {
-            Some(p) => p.0,
-            None => ast::EntityUID::unspecified_from_eid(ast::Eid::new("principal")),
-        };
-        let a = match action {
-            Some(a) => a.0,
-            None => ast::EntityUID::unspecified_from_eid(ast::Eid::new("action")),
-        };
-        let r = match resource {
-            Some(r) => r.0,
-            None => ast::EntityUID::unspecified_from_eid(ast::Eid::new("resource")),
-        };
-        Self(ast::Request::new(p, a, r, context.0))
-    }
-
-    /// Create a Request, validating that it complies with the given `Schema`.
-    ///
-    /// Note that you can create the `EntityUid`s using `.parse()` on any
-    /// string (via the `FromStr` implementation for `EntityUid`).
-    /// The principal, action, and resource fields are optional to support
-    /// the case where these fields do not contribute to authorization
-    /// decisions (e.g., because they are not used in your policies).
-    /// If any of the fields are `None`, we will automatically generate
-    /// a unique entity UID that is not equal to any UID in the store.
-    pub fn new_with_validation(
-        principal: Option<EntityUid>,
-        action: Option<EntityUid>,
-        resource: Option<EntityUid>,
-        context: Context,
-        schema: &Schema,
+        schema: Option<&Schema>,
     ) -> Result<Self, RequestValidationError> {
         let p = match principal {
             Some(p) => p.0,
@@ -3062,12 +3027,12 @@ impl Request {
             Some(r) => r.0,
             None => ast::EntityUID::unspecified_from_eid(ast::Eid::new("resource")),
         };
-        Ok(Self(ast::Request::new_with_validation(
+        Ok(Self(ast::Request::new(
             p,
             a,
             r,
             context.0,
-            &schema.0,
+            schema.map(|schema| &schema.0),
             Extensions::all_available(),
         )?))
     }
@@ -3124,8 +3089,7 @@ impl Context {
     /// Create an empty `Context`
     /// ```
     /// use cedar_policy::Context;
-    /// let c = Context::empty();
-    /// // let request: Request = Request::new(Some(principal), Some(action), Some(resource), c);
+    /// let context = Context::empty();
     /// ```
     pub fn empty() -> Self {
         Self(ast::Context::empty())
@@ -3136,7 +3100,7 @@ impl Context {
     /// of `(key, restricted expression)` pairs.
     /// ```
     /// use cedar_policy::{Context, RestrictedExpression};
-    /// use std::collections::HashMap;
+    /// # use std::collections::HashMap;
     /// use std::str::FromStr;
     /// # use cedar_policy::{Entities, EntityId, EntityTypeName, EntityUid, Request,PolicySet};
     /// let data : serde_json::Value = serde_json::json!({
@@ -3163,7 +3127,7 @@ impl Context {
     /// # let r_eid = EntityId::from_str("trip").unwrap();
     /// # let r_name: EntityTypeName = EntityTypeName::from_str("Album").unwrap();
     /// # let r = EntityUid::from_type_name_and_id(r_name, r_eid);
-    /// let request: Request = Request::new(Some(p), Some(a), Some(r), context);
+    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), context, None).unwrap();
     /// ```
     pub fn from_pairs(
         pairs: impl IntoIterator<Item = (String, RestrictedExpression)>,
@@ -3185,7 +3149,7 @@ impl Context {
     /// must specify the `Action` for schema-based parsing.
     /// ```
     /// use cedar_policy::{Context, RestrictedExpression};
-    /// use std::collections::HashMap;
+    /// # use std::collections::HashMap;
     /// use std::str::FromStr;
     /// # use cedar_policy::{Entities, EntityId, EntityTypeName, EntityUid, Request,PolicySet};
     /// let data =r#"{
@@ -3209,7 +3173,7 @@ impl Context {
     /// # let r_eid = EntityId::from_str("trip").unwrap();
     /// # let r_name: EntityTypeName = EntityTypeName::from_str("Album").unwrap();
     /// # let r = EntityUid::from_type_name_and_id(r_name, r_eid);
-    /// let request: Request = Request::new(Some(p), Some(a), Some(r), context);
+    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), context, None).unwrap();
     /// ```
     pub fn from_json_str(
         json: &str,
@@ -3236,33 +3200,36 @@ impl Context {
     /// must specify the `Action` for schema-based parsing.
     /// ```
     /// use cedar_policy::{Context, RestrictedExpression, Schema};
-    /// use std::collections::HashMap;
+    /// # use std::collections::HashMap;
     /// use std::str::FromStr;
     /// # use cedar_policy::{Entities, EntityId, EntityTypeName, EntityUid, Request,PolicySet};
     /// let data = serde_json::json!(
     /// {
-    ///     "sub": "1234"
+    ///     "sub": 1234
     /// });
-    /// let schema_data =r#"
+    /// let schema_json = serde_json::json!(
     ///     {
     ///       "": {
-    ///         "entityTypes": {},
-    ///           "actions": {
-    ///             "view": {
-    ///                "appliesTo": {
-    ///                  "principalTypes": [],
-    ///                   "resourceTypes": [],
-    ///                   "context": {
-    ///                     "type": "Record",
-    ///                     "attributes": {
-    ///                       "sub": { "type": "Long" }
-    ///                     }
+    ///         "entityTypes": {
+    ///           "User": {},
+    ///           "Album": {},
+    ///         },
+    ///         "actions": {
+    ///           "view": {
+    ///              "appliesTo": {
+    ///                "principalTypes": ["User"],
+    ///                 "resourceTypes": ["Album"],
+    ///                 "context": {
+    ///                   "type": "Record",
+    ///                   "attributes": {
+    ///                     "sub": { "type": "Long" }
     ///                   }
     ///                 }
     ///               }
+    ///             }
     ///           }
     ///       }
-    ///     }"#;
+    ///     });
     /// # // create a request
     /// # let p_eid = EntityId::from_str("alice").unwrap();
     /// # let p_name: EntityTypeName = EntityTypeName::from_str("User").unwrap();
@@ -3273,9 +3240,9 @@ impl Context {
     /// # let r_eid = EntityId::from_str("trip").unwrap();
     /// # let r_name: EntityTypeName = EntityTypeName::from_str("Album").unwrap();
     /// # let resource = EntityUid::from_type_name_and_id(r_name, r_eid);
-    /// let schema = Schema::from_str(schema_data).unwrap();
+    /// let schema = Schema::from_json_value(schema_json).unwrap();
     /// let context = Context::from_json_value(data, Some((&schema, &action))).unwrap();
-    /// let request: Request = Request::new(Some(principal), Some(action), Some(resource), context);
+    /// # let request: Request = Request::new(Some(principal), Some(action), Some(resource), context, Some(&schema)).unwrap();
     /// ```
     pub fn from_json_value(
         json: serde_json::Value,
@@ -3306,7 +3273,7 @@ impl Context {
     /// # use std::collections::HashMap;
     /// # use std::str::FromStr;
     /// # use std::fs::File;
-    /// let mut json = File::open("json_file.txt").expect("failed");
+    /// let mut json = File::open("json_file.txt").unwrap();
     /// let context = Context::from_json_file(&json, None).unwrap();
     /// # // create a request
     /// # let p_eid = EntityId::from_str("alice").unwrap();
@@ -3319,7 +3286,7 @@ impl Context {
     /// # let r_eid = EntityId::from_str("trip").unwrap();
     /// # let r_name: EntityTypeName = EntityTypeName::from_str("Album").unwrap();
     /// # let r = EntityUid::from_type_name_and_id(r_name, r_eid);
-    /// let request: Request = Request::new(Some(p), Some(a), Some(r), context);
+    /// # let request: Request = Request::new(Some(p), Some(a), Some(r), context, None).unwrap();
     /// ```
     pub fn from_json_file(
         json: impl std::io::Read,
@@ -3542,6 +3509,7 @@ mod partial_eval_test {
 #[cfg(test)]
 mod entity_uid_tests {
     use super::*;
+    use cool_asserts::assert_matches;
 
     /// building an `EntityUid` from components
     #[test]
@@ -3753,13 +3721,13 @@ permit(principal ==  A :: B
     #[test]
     fn accessing_unspecified_entity_returns_none() {
         let c = Context::empty();
-        let request: Request = Request::new(None, None, None, c);
+        let request = Request::new(None, None, None, c, None).unwrap();
         let p = request.principal();
         let a = request.action();
         let r = request.resource();
-        assert!(p.is_none());
-        assert!(a.is_none());
-        assert!(r.is_none());
+        assert_matches!(p, None);
+        assert_matches!(a, None);
+        assert_matches!(r, None);
     }
 }
 
@@ -4054,7 +4022,9 @@ mod policy_set_tests {
             Some(EntityUid::from_strs("Action", "a")),
             Some(EntityUid::from_strs("Resource", "b")),
             Context::empty(),
-        );
+            None,
+        )
+        .unwrap();
 
         let e = r#"[
             {
@@ -4336,7 +4306,9 @@ mod policy_set_tests {
             Some(EntityUid::from_strs("Action", "a")),
             Some(EntityUid::from_strs("Resource", "b")),
             Context::empty(),
-        );
+            None,
+        )
+        .unwrap();
 
         let e = r#"[
             {